6/17 13:24 下山　著者一覧の削除

+++ src/main/java/com/example/demo/controller/WriterController.java
@@ -3,6 +3,7 @@ package com.example.demo.controller;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,7 +12,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import com.example.demo.entity.Book;
 import com.example.demo.entity.Writer;
+import com.example.demo.repository.BookRepository;
 import com.example.demo.repository.WriterRepository;
 
 @Controller
@@ -19,6 +22,8 @@ import com.example.demo.repository.WriterRepository;
 public class WriterController {
 
 	private WriterRepository writerRepository;
+	@Autowired //←これ何？
+	private BookRepository bookRepository;
 
 	public WriterController(
 			WriterRepository writerRepository) {
@@ -28,6 +33,7 @@ public class WriterController {
 	//著者追加画面へ
 	@GetMapping("/writers")
 	public String index() {
+
 		return "WriterAdd";
 	}
 
@@ -123,6 +129,15 @@ public class WriterController {
 	public String deleteWriter(
 			@PathVariable Integer id,
 			Model model) {
+
+		List<Book> books = bookRepository.findByWriterId(id);
+
+		if (books.size() >= 1) {
+
+			model.addAttribute("message", "この著者は削除できません");
+			return "redirect:/admin/writers/list";
+		}
+
 		writerRepository.deleteById(id);
 
 		return "redirect:/admin/writers/list";
